Controle saisi

+++ src/app/components/apps/contacts/contacts.component.ts
@@ -1,13 +1,16 @@
-import { Component, OnInit, ViewChild, TemplateRef, ViewEncapsulation, } from '@angular/core';
-import { CalendarEvent, CalendarEventAction, CalendarEventTimesChangedEvent, CalendarView, } from 'angular-calendar';
-import { startOfDay, endOfDay, subDays, addDays, endOfMonth, isSameDay, isSameMonth, addHours, } from 'date-fns';
-import { Subject } from 'rxjs';
+import { Component, OnInit, ViewChild, TemplateRef, ViewEncapsulation, ElementRef } from '@angular/core';
+import { CalendarEvent, CalendarEventAction, CalendarEventTimesChangedEvent, CalendarView } from 'angular-calendar';
+import { startOfDay, endOfDay, subDays, addDays, endOfMonth, isSameDay, isSameMonth, addHours } from 'date-fns';
+import {Observable, Subject} from 'rxjs';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import 'flatpickr/dist/flatpickr.css';
 import PerfectScrollbar from 'perfect-scrollbar';
 import { ContactsService } from './contacts.service';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import {AbstractControl, FormBuilder, FormGroup, ValidationErrors, ValidatorFn, Validators} from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
+import { lettersOnlyValidator, phoneNumberValidator } from '../validators/custom-validators'; // Importez vos validateurs
+import {catchError, map, switchMap} from 'rxjs/operators';
+
 const colors: any = {
   red: {
     primary: '#705ec8',
@@ -27,83 +30,130 @@ const colors: any = {
   selector: 'app-contacts',
   templateUrl: './contacts.component.html',
   styleUrls: ['./contacts.component.scss'],
-  // encapsulation: ViewEncapsulation.None
-
+  encapsulation: ViewEncapsulation.None
 })
 export class ContactsComponent implements OnInit {
 
   coordonneeGroup: FormGroup;
-  contacts:any=[]
-  displayedContact:any={"nom":"","prenom":"","adresseSociete":"","commentaire":"","id":"","post":"","societe":"","telephone":"","telephoneSociete":"",}
-  loadContacts() {
-    this.contactService.getAllContacts().subscribe(
-      (data) => {
-        this.contacts = data; // Met à jour la liste des contacts dans le composant
-        console.log('Contacts:', this.contacts); // Affiche les contacts dans la console
-      },
-      (error) => {
-        console.error('Error fetching contacts:', error); // Gère les erreurs éventuelles
-      }
-    );
-  }
+  contacts: any[] = [];
+  filteredContacts: any[] = [];
+  displayedContact: any = {
+    nom: "",
+    prenom: "",
+    adresseSociete: "",
+    commentaire: "",
+    id: "",
+    post: "",
+    societe: "",
+    telephone: "",
+    telephoneSociete: "",
+    image: ''
+  };
+  searchText: string = '';
+  imageSrc: string = '../../assets/images/users/2.jpg'; // Default image source
+
+  @ViewChild('fileInput') fileInput: ElementRef;
+
   constructor(private modal: NgbModal,
               private contactService: ContactsService,
-              private toaster : ToastrService,
+              private toaster: ToastrService,
               private _formBuilder: FormBuilder) {}
 
   ngOnInit(): void {
     this.loadContacts();
-
-    this.displayedContact={"nom":"","prenom":"","adresseSociete":"","commentaire":"","id":"","post":"","societe":"","telephone":"","telephoneSociete":"",}
-    this.contacts;
-    this.allContacts()
-    this.emptyForm()
-
+    this.emptyForm();
   }
 
-  ngAfterViewInit(){
-    const scroll1 = document.querySelector('#mainContactList');
-
-    let ps = new PerfectScrollbar(scroll1);
-
+  onCameraClick(): void {
+    this.fileInput.nativeElement.click();
   }
 
- /* saveContact(){
+  /*onFileSelected(event: Event): void {
+    const input = event.target as HTMLInputElement;
+    if (input.files && input.files[0]) {
+      const file = input.files[0];
+      const reader = new FileReader();
 
-    if(this.coordonneeGroup.valid){
-      this.contactService.savecontact(
+      reader.onload = (e: any) => {
+        this.imageSrc = e.target.result;
+      };
 
-        this.coordonneeGroup.value
-
-      ).subscribe(data => {
-          this.toaster.success('Successfully Added')
-          this.allContacts()
-          this.closePopup()
-          this.emptyForm()
-
-
-
-        }
-
-
-      );
-
+      reader.readAsDataURL(file);
     }
-    else{
-      this.toaster.error('please, fill the form with valid data')
+  }
+*/
+  loadContacts() {
+    this.contactService.getAllContacts().subscribe(
+      (data) => {
+        this.contacts = data;
+        this.filteredContacts = this.contacts;
+        console.log('Contacts:', this.contacts);
+      },
+      (error) => {
+        console.error('Error fetching contacts:', error);
+      }
+    );
+  }
 
-    }
-  }*/
+  searchContacts() {
+    this.filteredContacts = this.contacts.filter(contact =>
+      (contact.fname + ' ' + contact.lname).toLowerCase().includes(this.searchText.trim().toLowerCase())
+    );
+  }
+
+  clearSearch() {
+    this.searchText = '';
+    this.filteredContacts = this.contacts;
+  }
   saveContact() {
     if (this.coordonneeGroup.valid) {
-      console.log('Valeur du formulaire:', this.coordonneeGroup.value); // Ajoutez cette ligne pour afficher la valeur du formulaire
+      const soctlph = this.coordonneeGroup.get('soctlph').value;
+      const tlph = this.coordonneeGroup.get('tlph').value;
+
+      this.contactService.checkPhoneExists(soctlph).pipe(
+        switchMap(soctlphExists => {
+          if (soctlphExists) {
+            this.toaster.error('Le numéro de téléphone de la société existe déjà.');
+            return new Observable<void>(); // Retourne un observable vide
+          } else {
+            return this.contactService.checkPhoneExists(tlph).pipe(
+              switchMap(tlphExists => {
+                if (tlphExists) {
+                  this.toaster.error('Le numéro de téléphone existe déjà.');
+                  return new Observable<void>(); // Retourne un observable vide
+                } else {
+                  return this.contactService.savecontact(this.coordonneeGroup.value).pipe(
+                    map(() => {
+                      this.toaster.success('Contact ajouté avec succès');
+                      this.coordonneeGroup.reset();
+                    }),
+                    catchError(error => {
+                      console.error('Erreur lors de l\'ajout du contact:', error);
+                      this.toaster.error('Erreur lors de l\'ajout du contact');
+                      return new Observable<void>();
+                    })
+                  );
+                }
+              })
+            );
+          }
+        })
+      ).subscribe();
+    } else {
+      this.toaster.error('Veuillez remplir le formulaire avec des données valides');
+    }
+  }
+  /*saveContact() {
+    if (this.coordonneeGroup.valid) {
+      console.log('Valeur du formulaire:', this.coordonneeGroup.value);
 
       this.contactService.savecontact(this.coordonneeGroup.value)
         .subscribe(
           data => {
             this.toaster.success('Contact ajouté avec succès');
-            this.closePopup(); // Fermer le popup ou modal
-            this.emptyForm(); // Vider le formulaire après ajout
+            this.closePopup();
+            this.emptyForm();
+            this.loadContacts();
           },
           error => {
             console.error('Erreur lors de l\'ajout du contact:', error);
@@ -113,54 +163,39 @@ export class ContactsComponent implements OnInit {
     } else {
       this.toaster.error('Veuillez remplir le formulaire avec des données valides');
     }
-  }
-
-  /*deleteContact(){
-    console.log("deleting : ",this.displayedContact)
-    this.contactService.deletecontact(this.displayedContact.id).subscribe(data =>
-      {
-        console.log("heloo delete")
-        // this.toaster.success('Successfully Deleted')
-        this.allContacts()
-      }    ,
-      err=>{
-        this.allContacts()
-
-      }
-
-
-    );
   }*/
-  deleteContact(id: any) {
-    console.log("Deleting contact with ID: ", id);
 
-    this.contactService.deletecontact(id).subscribe(
-      () => {
-        console.log("Delete successful");
-        // Ajoutez ici toute logique supplémentaire après la suppression réussie
-        this.allContacts(); // Exemple : rechargez la liste des contacts
-      },
-      error => {
-        console.error("Delete failed:", error);
-        // Gérez l'erreur ici (par exemple, affichez un message d'erreur à l'utilisateur)
-        this.allContacts(); // Exemple : rechargez la liste des contacts en cas d'erreur
-      }
-    );
+  deleteContact(id: any) {
+    const confirmation = confirm("Êtes-vous sûr de vouloir supprimer ce contact ?");
+    if (confirmation) {
+      this.contactService.deletecontact(id).subscribe(
+        () => {
+          console.log("Delete successful");
+          this.loadContacts();
+        },
+        error => {
+          console.error("Delete failed:", error);
+          this.loadContacts();
+        }
+      );
+    } else {
+      console.log("Suppression annulée par l'utilisateur");
+    }
   }
 
-  updateContact() {
+  updateContact(id:any) {
     if (this.coordonneeGroup.valid) {
-      const id = this.coordonneeGroup.value.id; // Récupérer l'ID depuis le formulaire
-
+      //const id = this.coordonneeGroup.value.id;
       this.contactService.updateContact(id, this.coordonneeGroup.value).subscribe(
         data => {
           this.toaster.success('Successfully updated');
-          this.closePopup(); // Exemple : fermer une popup après mise à jour réussie
-          this.emptyForm(); // Exemple : vider le formulaire après mise à jour réussie
+          this.closePopup();
+          this.emptyForm();
+          this.loadContacts();
         },
         error => {
           console.error('Update failed:', error);
-          this.toaster.error('Failed to update'); // Gérez l'erreur ici (par exemple, affichez un message d'erreur à l'utilisateur)
+          this.toaster.error('Failed to update');
         }
       );
     } else {
@@ -168,26 +203,9 @@ export class ContactsComponent implements OnInit {
     }
   }
 
-
-  allContacts(){
-
-    this.contactService.getAllContacts(
-    ).subscribe(data =>{
-        console.log("data = ", data)
-        this.contacts=data
-        if(data.length!=0){
-          this.displayedContact=data[0]
-        }
-      }
-    );
-
+  changeDisplayedContact(newDispContact: any) {
+    this.displayedContact = newDispContact;
   }
-
-  changeDisplayedContact(newDispContact){
-    this.displayedContact=newDispContact
-  }
-
-// popup
 
   @ViewChild('modalContent') modalContent!: TemplateRef<any>;
   modalRef: NgbModalRef | undefined;
@@ -199,53 +217,28 @@ export class ContactsComponent implements OnInit {
     event: null!
   };
 
+  isCreate: Boolean = false;
 
-  isCreate:Boolean=false
-
-  handleEvent(action: string, event: CalendarEvent, contact:any): void {
-
-    if(contact!=undefined && contact!=null){
-      this.isCreate=false
-      this.coordonneeGroup.patchValue(contact)
-    }else{
-      this.isCreate=true
+  handleEvent(action: string, event: CalendarEvent, contact: any): void {
+    if (contact != undefined && contact != null) {
+      this.isCreate = false;
+      this.coordonneeGroup.patchValue(contact);
+    } else {
+      this.isCreate = true;
     }
-    console.log("contact= ", contact)
+    console.log("contact= ", contact);
 
     this.modalData = { action, event };
     this.modalRef = this.modal.open(this.modalContent, { size: 'lg' });
   }
 
   closePopup(): void {
-    this.isCreate=false
-    this.emptyForm()
-
+    this.isCreate = false;
+    this.emptyForm();
     this.modalRef?.close();
   }
-  // Création du formulaire avec Reactive Form
-  /*id: ['',''],
-     nom: ['', Validators.required],
-     prenom: ['', Validators.required],
-     post: ['', Validators.required],
-     societe: ['', ''],
-     telephone: ['',  [Validators.required, Validators.pattern('^[0-9]+$')]],
-     telephoneSociete: ['',  [Validators.required, Validators.pattern('^[0-9]+$')]],
-     adresseSociete: ['', Validators.required],
-     commentaire: ['', ''],
-     id: ['',''],*/
-  /*emptyForm(){
-    this.coordonneeGroup = this._formBuilder.group({
-      fname: ['', Validators.required],
-      lname: ['', Validators.required],
-      societe: ['', Validators.required],
-      socadd: [''],
-      soctlph: ['', Validators.required],
-      tlph: [''],
-      post: [''],
-      comment: ['']
-    });
-  }*/
-  emptyForm() {
+
+ /* emptyForm() {
     this.coordonneeGroup = this._formBuilder.group({
       fname: [''],
       lname: [''],
@@ -256,23 +249,57 @@ export class ContactsComponent implements OnInit {
       post: [''],
       comment: ['']
     });
+  }*/
+  emptyForm() {
+    this.coordonneeGroup = this._formBuilder.group({
+      fname: ['', [Validators.required, lettersOnlyValidator()]],
+      lname: ['', [Validators.required, lettersOnlyValidator()]],
+      societe: [''],
+      socadd: [''],
+      soctlph: ['', [Validators.required, phoneNumberValidator()]],
+      tlph: ['', [Validators.required, phoneNumberValidator()]],
+      post: [''],
+      comment: ['']
+    });
   }
 
+  ngAfterViewInit() {
+    const scroll1 = document.querySelector('#mainContactList');
+    let ps = new PerfectScrollbar(scroll1);
+  }
+  // contacts.component.ts
+  onFileSelected(event: Event): void {
+    const input = event.target as HTMLInputElement;
+    if (input.files && input.files[0]) {
+      const file = input.files[0];
+      const reader = new FileReader();
 
+      reader.onload = (e: any) => {
+        this.imageSrc = e.target.result;
+      };
 
+      reader.readAsDataURL(file);
 
-  searchText: string = '';
-
-  get filteredContacts(): any[] {
-    if (!this.searchText.trim()) {
-      return this.contacts;
+      // Vérifiez que le contact est sélectionné
+      if (this.displayedContact && this.displayedContact.id) {
+        this.uploadImage(this.displayedContact.id, file);
+      } else {
+        this.toaster.error('Veuillez sélectionner un contact avant de télécharger une image.');
+      }
     }
-    return this.contacts.filter(contact =>
-      (contact.nom + ' ' + contact.prenom).toLowerCase().includes(this.searchText.trim().toLowerCase())
-    );
   }
 
-  clearSearch() {
-    this.searchText = '';
+  uploadImage(contactId: number, file: File): void {
+    this.contactService.uploadImage(contactId, file).subscribe(
+      (response) => {
+        this.displayedContact.image = response.image;
+        this.toaster.success('Image uploaded successfully');
+        this.loadContacts(); // Rechargez la liste des contacts pour refléter les modifications
+      },
+      (error) => {
+        console.error('Error uploading image:', error);
+        this.toaster.error('Error uploading image');
+      }
+    );
   }
 }
